avoid name conflict, remove duplicate word
