+++ Algorithmia/AlgorithmiaConf.swift
@@ -1,0 +1,13 @@
+//
+//  AlgorithmiaConf.swift
+//  AlgorithmiaSwiftDemo
+//
+//  Created by Erik Ilyin on 11/2/16.
+//  Copyright © 2016 algorithmia. All rights reserved.
+//
+
+import Foundation
+
+class Algo {
+    
+}
