+++ Algorithmia/Client/AlgoEntity.swift
@@ -10,9 +10,14 @@ import Foundation
 
 class AlgoEntity {
     
+    func headers() -> [AlgoRequest.HTTPHeader] {
+        return []
+    }
+    
     func body() -> Data? {
         return Data()
     }
+    
     
 }
 
@@ -23,7 +28,13 @@ class AlgoStringEntity:AlgoEntity {
         self.entity = entity
     }
     
+    override func headers() -> [AlgoRequest.HTTPHeader] {
+        return [AlgoRequest.HTTPHeader.ContentType(AlgoRequest.MIMEType.TEXT_PLAIN.rawValue)]
+    }
+    
     override func body() -> Data? {
         return entity.data(using: String.Encoding.utf8)
     }
 }
+
+
